voice-ai-poc/vs-voice-ai-backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^15.0"
fastapi = "^0.115.11"
twilio = "^9.4.6"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
langchain-community = "^0.3.19"
langgraph = "^0.3.5"
langchain-openai = "^0.3.8"
python-multipart = "^0.0.20"
noisereduce = "^3.0.0"
numpy = "^1.26.0"
scipy = "^1.13.0"
boto3 = "^1.35.0"
aws-sdk-bedrock-runtime = "*"


[build-system]
package-mode = false
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
